% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to_integer.R
\name{round_to_integer}
\alias{round_to_integer}
\title{Round to the nearest integer}
\usage{
round_to_integer(x)
}
\arguments{
\item{x}{A numeric element or vector to round to the nearest integer}
}
\value{
An integer element or vector
}
\description{
The [base::round()] function rounds the number `1.5` to `2` and the number
`2.5` also to `2`, because of the IEC 60559 standard. This function provides
an integer rounding alternative to [base::round()].
}
\examples{
library(dplyr)
# unexpected rounding
c(0.5, 1.5, 2.5, 3.5) |> round()
# rounding to nearest integer
c(0.5, 1.5, 2.5, 3.5) |> round_to_integer()

}
\author{
Gerko Vink \email{g.vink@uu.nl} and Hanne Oberman
\email{h.i.oberman@uu.nl}
}
